Use  elixir-pipes project to pipe functions while response is :ok
With pipe_matching/3 we can ask Elixir to pipe the value x from each step if it
matches {:ok, x}. We do so by basically converting each expression given to
case/2 as a step in the pipeline. As soon as any of the steps return something
that does not match {:ok, x}, the pipeline aborts, and returns the non-matching value.